Stefan: Fixed missing dependency due to remoteIP(). Fixed cmakelists
Changed 'uart.cpp' to 'fnc_uart.cpp' (name clash with uart.c from Driver/uart.)
Fixed include in C file.
Fixed signature of uart_read_bytes.
Fixed circular reference in components.

+++ FluidNC/include/Driver/fluidnc_uart.h
@@ -8,7 +8,7 @@ class InputPin;
 void uart_init(uint32_t uart_num);
 void uart_mode(uint32_t uart_num, uint32_t baud, UartData dataBits, UartParity parity, UartStop stopBits);
 bool uart_half_duplex(uint32_t uart_num);
-int  uart_read(uint32_t uart_num, uint8_t* buf, size_t len, uint32_t timeout_ms);
+int  uart_read(uint32_t uart_num, uint8_t* buf, uint32_t len, uint32_t timeout_ms);
 int  uart_write(uint32_t uart_num, const uint8_t* buf, size_t len);
 void uart_xon(uint32_t uart_num);
 void uart_xoff(uint32_t uart_num);
